+++ CHANGELOG.md
@@ -5,7 +5,9 @@
 * Bean Searcher
   * 新增条件分组 与 组逻辑运算功能（TODO）
   * 新增 `EnumFieldConvertor`，用来做枚举字段转换
-  * SearchBean 映射字段支持从父类继承（TODO）
+  * SearchBean 的映射字段支持从父类继承
+  * 注解 @SearchBean 也支持从父类继承
+  * 支持非静态内部类形式的 SearchBean
 * Bean Searcher Boot Starter
   * 默认自动配置添加 `EnumFieldConvertor`